+++ lib/startback/bus/bunny.rb
@@ -1,0 +1,1 @@
+require_relative 'bunny/async'