+++ spec/spec_helper.rb
@@ -1,5 +1,6 @@
 $LOAD_PATH.unshift File.expand_path('../../lib', __FILE__)
 require 'startback'
+require 'startback/bus'
 require 'rack/test'
 
 module SpecHelpers